requirements.txt
PyQt6==6.5.0
Pillow==9.5.0
pyobjc-framework-Quartz==9.0.1
